Make rotate-credential and discard-old-credential top-level commands
Review feedback: turning "kubectl moco credential" into a parent
command with rotate/discard/show subcommands breaks backward
compatibility — cobra resolves subcommands before positional args, so
clusters literally named "rotate" or "discard" would trigger a rotation
instead of showing their credential. The fallback-to-show behavior
cannot avoid that collision, and the subcommand plumbing was growing
hard to maintain.

Restore "kubectl moco credential" to its released single-purpose form
and expose the rotation operations as top-level commands instead:

* kubectl moco rotate-credential CLUSTER_NAME
* kubectl moco discard-old-credential CLUSTER_NAME

The command logic (stale-CR detection, idle/awaiting-discard guards,
generation patching) is unchanged, only relocated. Update
docs/kubectl-moco.md and the design doc accordingly; the doc's stale
"Phase" wording is replaced with the condition-based description while
here.
